requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
psycopg2>=2.9.9
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.10
langgraph>=0.0.10
groq>=0.4.2
httpx>=0.26.0


